ValidateItem类更改为ValidateRule
